listings_scraper/requirements.txt
scrapy==2.10.1
psycopg2-binary==2.9.9